+++ externals/comin/cmake/Scripts/predef_replay_data.cmake
@@ -1,6 +1,0 @@
-comin_add_replay_data(NAME test_nwp_R02B04
-                      URL https://gitlab.dkrz.de/api/v4/projects/139706/packages/generic/replay_data/1/42a9f86133e6b98a25937036531b6494_test_nwp_R02B04.tar.gz
-                      MD5HASH 42a9f86133e6b98a25937036531b6494)
-comin_add_replay_data(NAME test_nwp_R02B04_R02B05_nest
-                      URL https://gitlab.dkrz.de/api/v4/projects/139706/packages/generic/replay_data/1/5a00067fa51e9d56ca00a3c1f5ae2342_test_nwp_R02B04_R02B05_nest.tar.gz
-                      MD5HASH 5a00067fa51e9d56ca00a3c1f5ae2342)