+++ externals/comin/plugins/python_adapter/test/test_print.py
@@ -1,5 +1,0 @@
-import comin
-
-comin.print_debug("This is a debug message!")
-comin.print_info("This is a info message!")
-comin.print_warning("This is a warning!")