+++ externals/comin/plugins/python_adapter/test/uncaught_exception.py
@@ -1,6 +1,0 @@
-import comin
-
-
-@comin.EP_SECONDARY_CONSTRUCTOR
-def sec_ctr():
-    raise RuntimeError("FOOO")